Data used for titanic survival prediction
'train.csv' will contain the details of a subset of the passengers on board (891 to be exact) and importantly, will reveal whether they survived or not, also known as the “ground truth”.

The `test.csv` dataset contains similar information but does not disclose the “ground truth” for each passenger. It’s your job to predict these outcomes.

'gender_submission.csv' a set of predictions that assume all and only female passengers survive, as an example of what a submission file should look like.